GrapheneQFT.jl

GrapheneQFT — Module

This package provides provides a set of functions to facilitate the field-theoretic treatment of monolayer graphene using the tight-binding model. The Hamiltonian employed by this package includes only the nearest-neighbor hopping term with $t = 2.8$eV. The derivation of the formalism is available here.

Example:

Pick a few graphene coordinates:

a1 = graphene_A(0, 17)
a2 = graphene_A(1, 1)
a3 = graphene_B(3, -2)
a4 = graphene_A(2, 7)
a5 = graphene_A(0, 1)

Next, initialize a couple of impurities and couple them to the coordinates:

ϵ1 = 1.7
ϵ2 = -0.2

V1 = 0.4
V2 = 2+0.9im
V3 = -2.1

imp1 = ImpurityState(ϵ1, [Coupling(V1, a4), Coupling(V2, a3)])
imp2 = ImpurityState(ϵ2, [Coupling(V3, a5)])

Create a few coupling integrals c and generate a coupling dictionary pert:

c1 = ComplexF64(3)
c2 = ComplexF64(5 + 1im)
c3 = ComplexF64(9 - 2im)

pert = new_perturbation()

Add the c's to the dictionary and initialize the system:

pert = add_perturbation(pert, a1, a1, c1)
pert = add_perturbation(pert, a2, a1, c2)
pert = add_perturbation(pert, a3, a1, c3)

μ = 0.2
T = 0.0

my_system = mk_GrapheneSystem(μ, T, [imp1, imp2], pert)

my_system can now be used to calculate the relevant Green's functions (δG_R, G_R, δΓ, Γ), from which quantities of interest can be computed.

Module Index

GrapheneQFT.Coupling

GrapheneQFT.GrapheneCoord

GrapheneQFT.GrapheneSystem

GrapheneQFT.ImpurityState

GrapheneQFT.Location

GrapheneQFT.Perturbation

GrapheneQFT.G_R

GrapheneQFT.add_perturbation

GrapheneQFT.graphene_A

GrapheneQFT.graphene_B

GrapheneQFT.mk_GrapheneSystem

GrapheneQFT.new_perturbation

GrapheneQFT.Γ

GrapheneQFT.δG_R

GrapheneQFT.δΓ

Detailed API

GrapheneQFT.Coupling — Type

struct Coupling
    V::ComplexF64           # Coupling to a graphene atom
    coord::GrapheneCoord    # Location of the graphene atom
end

A structure describing the coupling V (in eV) between an impurity state and a graphene atom at coord.

GrapheneQFT.GrapheneCoord — Type

struct GrapheneCoord
    u::Int
    v::Int
    sublattice::Sublattice
end

Lattice coordinate of a carbon atom, generated using graphene_A or graphene_B. Each coordinate contains the sublattice index, as well as the integer coefficients of the two basis vectors $d\times(\pm 1 \hat{x} + \sqrt{3}\hat{y}) / 2$, where $d = 2.46$Å is the graphene lattice constant.

GrapheneQFT.GrapheneSystem — Type

struct GrapheneSystem
    μ::Float64                              # Chemical potential
    T::Float64                              # Temperature
    Δ::Array{ComplexF64,2}                  # Δ matrix
    V::Array{ComplexF64,2}                  # V Matrix
    scattering_atoms::Vector{GrapheneCoord} # List of all perturbed atoms
    imps::Vector{Float64}                   # Impurity energies
end

A structure describing the perturbed graphene system. See mk_GrapheneSystem for details.

GrapheneQFT.ImpurityState — Type

struct ImpurityState
    ϵ::Float64                  # Impurity state energy
    coupling::Vector{Coupling}  # Coupling array for the impurity
end

A structure describing an impurity state with energy ϵ (in eV) and containing a list of all its couplings to graphene atoms.

GrapheneQFT.Location — Type

struct Location
    x::Float64
    y::Float64
    z::Float64
end

A structure describing a point in 3D space, with the lengths in Å

GrapheneQFT.Perturbation — Type

struct Perturbation
    pert::Dict{Tuple{GrapheneCoord,GrapheneCoord},ComplexF64}
end

A structure wrapping a dictionary describing the coupling between graphene atoms and local potential. See new_perturbation() and add_perturbation for more details.

GrapheneQFT.G_R — Method

G_R(z, a1::GrapheneCoord, a2::GrapheneCoord, s::GrapheneSystem)

The full real-space graphene Green's function in the presence of defects as a function of complex energy z.

Arguments

z: complex energy

a1: GrapheneCoord of the first atom

a2: GrapheneCoord of the second atom

s: GrapheneSystem for which G_R is calculated

Output

ComplexF64

GrapheneQFT.add_perturbation — Method

function add_perturbation(
    p::Perturbation,
    a::GrapheneCoord,
    b::GrapheneCoord,
    Δ::ComplexF64,
)

A function that adds coupling between atoms a and b or, if a == b, an on-site potential to an existing Perturbation structure. When adding a coupling between two atoms, two entries are added to the dictionary: one (a,b)=>Δ and another one (b,a)=>conj(Δ). Passing the same pair more than once overwrites the previous entries.

Arguments

p: Perturbation to which an additional element is to be added

a: GrapheneCoord of the first atom

b: GrapheneCoord of the second atom

Δ: coupling between the atoms

Output

Perturbation with the new coupling added

GrapheneQFT.graphene_A — Method

graphene_A(u::Int, v::Int)

Create a GrapheneCoord for an atom belonging to sublattice A at the unit cell (u, v)

Arguments

u: coefficient of basis vector $d\times(1 \hat{x} + \sqrt{3}\hat{y}) / 2$

v: coefficient of basis vector $d\times(-1 \hat{x} + \sqrt{3}\hat{y}) / 2$

Output

GrapheneCoord of the carbon atom

GrapheneQFT.graphene_B — Method

graphene_B(u::Int, v::Int)

Create a GrapheneCoord for an atom belonging to sublattice B at the unit cell (u, v)

Arguments

u: coefficient of basis vector $d\times(1 \hat{x} + \sqrt{3}\hat{y}) / 2$

v: coefficient of basis vector $d\times(-1 \hat{x} + \sqrt{3}\hat{y}) / 2$

Output

GrapheneCoord of the carbon atom

GrapheneQFT.mk_GrapheneSystem — Method

function mk_GrapheneSystem(
    μ::Float64,
    T::Float64,
    imps::Vector{ImpurityState},
    p::Perturbation,
)

A function for constructing GrapheneSystem.

Arguments

μ: chemical potential

T: Temperature

imps: a list of [ImpurityState]'s

p: Perturbation struct containing the carbon-carbon coupling

Output

GrapheneSystem with the Δ and V matrix computed from imps and p.

In addition, a list of all GrapheneCoord that are perturbed and a lise of impurity energies are included.

GrapheneQFT.new_perturbation — Method

new_perturbation()

A helper initialization function , giving an empty dictionary wrapped in the Perturbation structure. The keys for the dictionary are tuples of two GrapheneCoord, while the values are ComplexF64.

Arguments:

None

Output:

Perturbation wrapping an empty Dict{Tuple{GrapheneCoord,GrapheneCoord},ComplexF64}

GrapheneQFT.Γ — Method

Γ(z, s::GrapheneSystem)

The full impurity Green's function with the correction due to the impurities' interaction with graphene.

Arguments

z: complex energy

s: GrapheneSystem for which Γ_R is calculated

Output

Matrix{ComplexF64}

GrapheneQFT.δG_R — Method

δG_R(z, a1::GrapheneCoord, a2::GrapheneCoord, s::GrapheneSystem)

The correction to the real-space graphene Green's function in the presence of defects as a function of complex energy z.

Arguments

z: complex energy

a1: GrapheneCoord of the first atom

a2: GrapheneCoord of the second atom

s: GrapheneSystem for which δG_R is calculated

Output

ComplexF64

GrapheneQFT.δΓ — Method

δΓ(z, s::GrapheneSystem)

The correction to the impurity Green's function due to the impurities' interaction with graphene.

Arguments

z: complex energy

s: GrapheneSystem for which δΓ_R is calculated

Output

Matrix{ComplexF64}