The GrapheneQFT Module

GrapheneQFT — Module

GrapheneQFT

$t = 2.8$eV

Module Index

GrapheneQFT.GrapheneCoord

GrapheneQFT.Location

GrapheneQFT.crystal_to_cartesian

GrapheneQFT.graphene_A

GrapheneQFT.graphene_B

GrapheneQFT.graphene_neighbors

GrapheneQFT.graphene_propagator

GrapheneQFT.propagator_matrix

Detailed API

GrapheneQFT.GrapheneCoord — Type

Lattice coordinate of a carbon atom, generated using graphene_A or graphene_B. Each coordinate contains the sublattice index, as well as the integer coefficients of the two basis vectors $d\times(\pm 1 \hat{x} + \sqrt{3}\hat{y}) / 2$, where $d = 2.46$Å is the graphene lattice constant.

GrapheneQFT.Location — Type

Location(x::Float64, y::Float64, z::Float64)

A structure describing a point in 3D space, with the lengths in Å

GrapheneQFT.crystal_to_cartesian — Method

crystal_to_cartesian(coord::GrapheneCoord)

Convert a crystal coordinate to a cartesian one.

Arguments

coord: a GrapheneCoord that needs to be converted to a 3D Location

Output

Location of the carbon atom

GrapheneQFT.graphene_A — Method

graphene_A(u::Int, v::Int)

Create a GrapheneCoord for an atom belonging to sublattice A at the unit cell (u, v)

Arguments

u: coefficient of basis vector $d\times(1 \hat{x} + \sqrt{3}\hat{y}) / 2$

v: coefficient of basis vector $d\times(-1 \hat{x} + \sqrt{3}\hat{y}) / 2$

Output

GrapheneCoord of the carbon atom

GrapheneQFT.graphene_B — Method

graphene_B(u::Int, v::Int)

Create a GrapheneCoord for an atom belonging to sublattice B at the unit cell (u, v)

Arguments

u: coefficient of basis vector $d\times(1 \hat{x} + \sqrt{3}\hat{y}) / 2$

v: coefficient of basis vector $d\times(-1 \hat{x} + \sqrt{3}\hat{y}) / 2$

Output

GrapheneCoord of the carbon atom

GrapheneQFT.graphene_neighbors — Method

graphene_neighbors(atom::GrapheneCoord)

Determine the nearest neighbors of an atom.

Arguments

atom: GrapheneCoord whose nearest neighbors are to be determined

Output

A Vector of GrapheneCoord containing the neighbors of atom

GrapheneQFT.graphene_propagator — Method

graphene_propagator(atom1::GrapheneCoord, atom2::GrapheneCoord, z)

The propagator function picks out the correct element of the Ξ matrix based on the sublattices of the graphene coordinates.

GrapheneQFT.propagator_matrix — Method

graphene_propagator_matrix(z, coords::Vector{GrapheneCoord})

Given a list of GrapheneCoord, this functiohn returns a Ξ(z) propagator matrix. The calculation is sped up using the fact that the matrix is symmetric.