GrapheneQFT.jl

GrapheneQFT — Module

This package provides provides a set of functions to facilitate the field-theoretic treatment of monolayer graphene using the tight-binding model. The Hamiltonian employed by this package includes only the nearest-neighbor hopping term with $t = 2.8$eV. The derivation of the formalism is available here.

Module Index

GrapheneQFT.Coupling

GrapheneQFT.GrapheneCoord

GrapheneQFT.GrapheneSystem

GrapheneQFT.ImpurityState

GrapheneQFT.Location

GrapheneQFT.G_R

GrapheneQFT.add_coupling!

GrapheneQFT.add_impurity!

GrapheneQFT.add_perturbation!

GrapheneQFT.crystal_to_cartesian

GrapheneQFT.graphene_A

GrapheneQFT.graphene_B

GrapheneQFT.graphene_neighbors

GrapheneQFT.new_graphene_system

GrapheneQFT.new_impurity

GrapheneQFT.remove_coupling!

GrapheneQFT.remove_impurity!

GrapheneQFT.remove_perturbation!

GrapheneQFT.set_T!

GrapheneQFT.set_μ!

GrapheneQFT.Γ

GrapheneQFT.δG_R

GrapheneQFT.δΓ

Detailed API

GrapheneQFT.Coupling — Type

struct Coupling
    V::Float64              # Coupling to a graphene atom
    coord::GrapheneCoord    # Location of the graphene atom
end

A structure describing the coupling V (in eV) between an impurity state and a graphene atom at coord.

GrapheneQFT.GrapheneCoord — Type

struct GrapheneCoord
    u::Int
    v::Int
    sublattice::Sublattice
end

Lattice coordinate of a carbon atom, generated using graphene_A or graphene_B. Each coordinate contains the sublattice index, as well as the integer coefficients of the two basis vectors $d\times(\pm 1 \hat{x} + \sqrt{3}\hat{y}) / 2$, where $d = 2.46$Å is the graphene lattice constant.

GrapheneQFT.GrapheneSystem — Type

mutable struct GrapheneSystem
    μ::Float64                          # Chemical potential
    T::Float64                          # Temperature
    imps::Vector{ImpurityState}         # Impurity states in the system
    pert::Dict{Tuple{GrapheneCoord,GrapheneCoord},Float64}  # Direct perturbation
    Δ::Array{Float64,2}                 # Δ matrix
    V::Array{Float64,2}                 # V Matrix
    scattering_atoms::Vector{GrapheneCoord} # List of all perturbed atoms
end

A structure describing the perturbed graphene system. Whenever an impurity or a direct perturbation (coupling between graphene atoms) is added or removed, the Δ and V matrices are updated, as is the list of the perturbed atoms scattering_atoms.

GrapheneQFT.ImpurityState — Type

mutable struct ImpurityState
    ϵ::Float64                  # Impurity state energy
    coupling::Vector{Coupling}  # Coupling array for the impurity
end

A structure describing an impurity state with energy ϵ (in eV) and containing a list of all its couplings to graphene atoms.

GrapheneQFT.Location — Type

struct Location
    x::Float64
    y::Float64
    z::Float64
end

A structure describing a point in 3D space, with the lengths in Å

GrapheneQFT.G_R — Method

G_R(z, a1::GrapheneCoord, a2::GrapheneCoord, s::GrapheneSystem)

The full real-space graphene Green's function in the presence of defects as a function of complex energy z.

Arguments

z: complex energy

a1: GrapheneCoord of the first atom

a2: GrapheneCoord of the second atom

s: GrapheneSystem for which G_R is calculated

Output

ComplexF64

GrapheneQFT.add_coupling! — Method

add_coupling!(imp::ImpurityState, V::Float64, coord::GrapheneCoord)

A function to add a coupling to an existing ImpurityState.

Arguments

imp: an existing ImpurityState

V: coupling energy (in eV)

coord: GrapheneCoord to which one wishes to couple the impurity

Output

ImpurityState with an added element in the coupling list

GrapheneQFT.add_impurity! — Method

add_impurity!(s::GrapheneSystem, imp::ImpurityState)

A function to add an impurity to an existing GrapheneSystem.

Arguments

s: an existing GrapheneSystem

imp: the ImpurityState to be added

Output

GrapheneSystem with an added element in the impurity list

GrapheneQFT.add_perturbation! — Method

add_perturbation!(
    s::GrapheneSystem,
    a::GrapheneCoord,
    b::GrapheneCoord,
    Δ::Float64,
)

Introduce a direct coupling between graphene atoms a and b or, if a==b, a local potential

Arguments

s: an existing GrapheneSystem

a: GrapheneCoord of the first atom

b: GrapheneCoord of the second atom

Δ: coupling between the atoms

Output

GrapheneSystem with the newly-added coupling

GrapheneQFT.crystal_to_cartesian — Method

crystal_to_cartesian(coord::GrapheneCoord)

Convert a crystal coordinate to a cartesian one for plotting.

Arguments

coord: a GrapheneCoord that needs to be converted to a 3D Location

Output

Location of the carbon atom

GrapheneQFT.graphene_A — Method

graphene_A(u::Int, v::Int)

Create a GrapheneCoord for an atom belonging to sublattice A at the unit cell (u, v)

Arguments

u: coefficient of basis vector $d\times(1 \hat{x} + \sqrt{3}\hat{y}) / 2$

v: coefficient of basis vector $d\times(-1 \hat{x} + \sqrt{3}\hat{y}) / 2$

Output

GrapheneCoord of the carbon atom

GrapheneQFT.graphene_B — Method

graphene_B(u::Int, v::Int)

Create a GrapheneCoord for an atom belonging to sublattice B at the unit cell (u, v)

Arguments

u: coefficient of basis vector $d\times(1 \hat{x} + \sqrt{3}\hat{y}) / 2$

v: coefficient of basis vector $d\times(-1 \hat{x} + \sqrt{3}\hat{y}) / 2$

Output

GrapheneCoord of the carbon atom

GrapheneQFT.graphene_neighbors — Method

graphene_neighbors(atom::GrapheneCoord)

Determine the nearest neighbors of an atom.

Arguments

atom: GrapheneCoord whose nearest neighbors are to be determined

Output

A Vector of GrapheneCoord containing the neighbors of atom

GrapheneQFT.new_graphene_system — Method

new_graphene_system()

Create a new GrapheneSystem.

Arguments

None

Output

GrapheneSystem with $T = 0$, $μ = 0$, and no defects.

GrapheneQFT.new_impurity — Method

new_impurity(ϵ::Float64)

A function to create a new impurity.

Arguments

ϵ: energy of the impurity state

Output

ImpurityState with an empty coupling list

GrapheneQFT.remove_coupling! — Method

remove_coupling!(imp::ImpurityState, ind::Int)

A function to remove a coupling from an existing ImpurityState.

Arguments

imp: an existing ImpurityState

ind: the index of the coupling in the coupling list to be removed

Output

ImpurityState with the desired coupling removed

GrapheneQFT.remove_impurity! — Method

remove_impurity!(s::GrapheneSystem, ind::Int)

A function to remove an impurity from an existing GrapheneSystem.

Arguments

s: an existing GrapheneSystem

ind: the index of the impurity in the impurity list to be removed

Output

GrapheneSystem with the desired impurity removed

GrapheneQFT.remove_perturbation! — Method

remove_perturbation!(
    s::GrapheneSystem,
    a::GrapheneCoord,
    b::GrapheneCoord,
)

Remove a coupling between graphene atoms a and b or, if a==b, a local potential

Arguments

s: an existing GrapheneSystem

a: GrapheneCoord of the first atom

b: GrapheneCoord of the second atom

Output

GrapheneSystem with the coupling removed

GrapheneQFT.set_T! — Method

set_T!(s::GrapheneSystem, new_T::Float64)

Modify the temperature of an existing GrapheneSystem

Arguments

s: an existing GrapheneSystem

new_T: new temperature

Output

GrapheneSystem with the new temperature new_T

GrapheneQFT.set_μ! — Method

set_μ!(s::GrapheneSystem, new_μ::Float64)

Modify the chemical potential of an existing GrapheneSystem

Arguments

s: an existing GrapheneSystem

new_μ: new chemical potential

Output

GrapheneSystem with the new chemical potential new_μ

GrapheneQFT.Γ — Method

Γ(z, s::GrapheneSystem)

The full impurity Green's function with the correction due to the impurities' interaction with graphene.

Arguments

z: complex energy

s: GrapheneSystem for which Γ_R is calculated

Output

Matrix{ComplexF64}

GrapheneQFT.δG_R — Method

δG_R(z, a1::GrapheneCoord, a2::GrapheneCoord, s::GrapheneSystem)

The correction to the real-space graphene Green's function in the presence of defects as a function of complex energy z.

Arguments

z: complex energy

a1: GrapheneCoord of the first atom

a2: GrapheneCoord of the second atom

s: GrapheneSystem for which δG_R is calculated

Output

ComplexF64

GrapheneQFT.δΓ — Method

δΓ(z, s::GrapheneSystem)

The correction to the impurity Green's function due to the impurities' interaction with graphene.

Arguments

z: complex energy

s: GrapheneSystem for which δΓ_R is calculated

Output

Matrix{ComplexF64}