API

GrapheneQFT.GrapheneCoord

GrapheneQFT.GrapheneSystem

GrapheneQFT.ImpurityState

GrapheneQFT.G_R

GrapheneQFT.graphene_A

GrapheneQFT.graphene_B

GrapheneQFT.mkGrapheneSystem

GrapheneQFT.Γ

GrapheneQFT.δG_R

GrapheneQFT.δΓ

Defect-Free Graphene

GrapheneQFT.GrapheneCoord — Type

GrapheneCoord(u::Int, v::Int, sublattice::Sublattice)

Lattice coordinate of a carbon atom, generated using graphene_A or graphene_B.

Each coordinate contains the sublattice index A or B, as well as the integer coefficients of the two basis vectors $d\times(\pm 1 \hat{x} + \sqrt{3}\hat{y}) / 2$ (u for +, v for -), with $d = 2.46$Å as the lattice constant.

GrapheneQFT.graphene_A — Function

graphene_A(u::Int, v::Int)

Create a GrapheneCoord for an atom belonging to sublattice A at the unit cell (u, v).

GrapheneQFT.graphene_B — Function

graphene_B(u::Int, v::Int)

Create a GrapheneCoord for an atom belonging to sublattice B at the unit cell (u, v).

Defect Manipulation

GrapheneQFT.ImpurityState — Type

ImpurityState(ϵ::Float64, coupling::Vector{Tuple{Float64,GrapheneCoord}})

An impurity state of energy ϵ (in eV) coupled to the graphene system. The tuples in the coupling field contain all the coupling energies (in eV) and the corresponding GrapheneCoord's.

GrapheneQFT.GrapheneSystem — Type

GrapheneSystem(
    μ::Float64,
    T::Float64,
    Δ::Array{Float64,2},
    V::Array{Float64,2},
    scattering_atoms::Vector{GrapheneCoord},
    imps::Vector{Float64},
)

A structure describing the perturbed graphene system.

See also mkGrapheneSystem.

GrapheneQFT.mkGrapheneSystem — Function

mkGrapheneSystem(
    μ::Float64,
    T::Float64,
    imps::Vector{ImpurityState},
    pert::Vector{Tuple{GrapheneCoord,GrapheneCoord,Float64}},
)

Construct GrapheneSystem.

Arguments

μ: chemical potential

T: temperature

imps: a list of [ImpurityState]'s

pert: a list of 3-tuples describing the coupling between carbon atoms

When supplying pert, ensure that each coordinate pair appears only once as repeated pairs with different couplings will cause eariler values to be overwritten. The order of the coordinates does not matter. A tuple with the same coordinate given twice creates and on-site potential for the coordinate.

The function constructs a GrapheneSystem with the Δ and V matrix computed from imps and pert. In addition, a list of all GrapheneCoord that are perturbed (scattering_atoms field in GrapheneCoord) and a list of impurity energies (imps field in GrapheneCoord) are included.

Green's Functions

GrapheneQFT.δG_R — Function

δG_R(z, a1::GrapheneCoord, a2::GrapheneCoord, s::GrapheneSystem)

The correction to the real-space graphene Green's function in the presence of defects as a function of complex energy z.

Arguments

z: complex energy

a1: GrapheneCoord of the first atom

a2: GrapheneCoord of the second atom

s: GrapheneSystem for which δG_R is calculated

GrapheneQFT.G_R — Function

G_R(z, a1::GrapheneCoord, a2::GrapheneCoord, s::GrapheneSystem)

The full real-space graphene Green's function in the presence of defects as a function of complex energy z.

Arguments

z: complex energy

a1: GrapheneCoord of the first atom

a2: GrapheneCoord of the second atom

s: GrapheneSystem for which G_R is calculated

GrapheneQFT.δΓ — Function

δΓ(z, s::GrapheneSystem)

The correction to the impurity Green's function due to the impurities' interaction with graphene.

Arguments

z: complex energy

s: GrapheneSystem for which δΓ is calculated

GrapheneQFT.Γ — Function

Γ(z, s::GrapheneSystem)

The full impurity Green's function with the correction due to the impurities' interaction with graphene.

Arguments

z: complex energy

s: GrapheneSystem for which Γ is calculated