API

GrapheneQFT.noimps

GrapheneQFT.nopert

GrapheneQFT.GrapheneCoord

GrapheneQFT.GrapheneSystem

GrapheneQFT.ImpurityState

GrapheneQFT.G_R

GrapheneQFT.coulomb_energy_pz_pz

GrapheneQFT.coulomb_potential_pz

GrapheneQFT.coulomb_potential_pz_interp

GrapheneQFT.crystal_to_cartesian

GrapheneQFT.graphene_multiple_neighbors

GrapheneQFT.graphene_neighbors

GrapheneQFT.mkGrapheneSystem

GrapheneQFT.peierls_phase

GrapheneQFT.Γ

GrapheneQFT.Ψ_pz

GrapheneQFT.δG_R

GrapheneQFT.δΓ

Pristine Graphene

GrapheneQFT.GrapheneCoord — Type

GrapheneCoord(u::Int, v::Int, sublattice::Sublattice)

Lattice coordinate of a carbon atom.

Each coordinate contains the sublattice index A or B, as well as the integer coefficients of the two basis vectors $d\times(\pm 1 \hat{x} + \sqrt{3}\hat{y}) / 2$ (u for +, v for -), with $d = 2.46$Å as the lattice constant.

GrapheneQFT.ImpurityState — Type

ImpurityState(ϵ::Float64, coupling::Vector{Tuple{Float64,GrapheneCoord}})

An impurity state of energy ϵ (in eV) coupled to the graphene system. The tuples in the coupling field contain all the coupling energies (in eV) and the corresponding GrapheneCoord's.

GrapheneQFT.GrapheneSystem — Type

GrapheneSystem(
    μ::Float64,
    T::Float64,
    Δ::Array{ComplexF64,2},
    V::Array{Float64,2},
    scattering_atoms::Vector{GrapheneCoord},
    imps::Vector{Float64},
)

A structure describing the perturbed graphene system.

See also mkGrapheneSystem.

GrapheneQFT.graphene_neighbors — Function

graphene_neighbors(atom::GrapheneCoord)

Determine the nearest neighbors of an atom and return a vector of the correspoding GrapheneCoord's.

GrapheneQFT.graphene_multiple_neighbors — Function

graphene_multiple_neighbors(atom::GrapheneCoord, n::Int)

Return an array of GrapheneCoord's obtained by iteratively running graphene_neighbors n times, applying it only to the newly-added GrapheneCoord's from the past iteration. The entries in the result are unique and sorted by their distance from atom.

Arguments

atom: GrapheneCoord from which the iteration begins

n: number of iterations of graphene_neighbors.

GrapheneQFT.crystal_to_cartesian — Function

crystal_to_cartesian(coord::GrapheneCoord)

Convert a GrapheneCoord to a cartesian point with lengths in Å, where GrapheneCoord(0, 0, A) is at the origin.

GrapheneQFT.noimps — Constant

const noimps = ImpurityState[]

An empty array to be used in constructing the GrapheneSystem if there are no impurity states.

GrapheneQFT.nopert — Constant

const nopert = Tuple{GrapheneCoord,GrapheneCoord,ComplexF64}[]

An empty array to be used in constructing the GrapheneSystem if there are no direct perturbation terms.

GrapheneQFT.mkGrapheneSystem — Function

mkGrapheneSystem(
    μ::Float64,
    T::Float64,
    imps::Vector{ImpurityState},
    pert::Vector{Tuple{GrapheneCoord,GrapheneCoord,ComplexF64}},
)

Construct GrapheneSystem.

Arguments

μ: chemical potential

T: temperature

imps: a list of [ImpurityState]'s

pert: a list of 3-tuples describing the coupling between carbon atoms

When supplying pert, ensure that each coordinate pair appears only once as repeated pairs with different couplings will cause eariler values to be overwritten. The order of the coordinates does not matter. A tuple with the same coordinate given twice creates and on-site potential for the coordinate.

The function constructs a GrapheneSystem with the Δ and V matrix computed from imps and pert. In addition, a list of all GrapheneCoord that are perturbed (scattering_atoms field in GrapheneCoord) and a list of impurity energies (imps field in GrapheneCoord) are included.

GrapheneQFT.peierls_phase — Function

peierls_phase(vec_pot, a1::GrapheneCoord, a2::GrapheneCoord)

Calculate the phase used in the Peierls substitution to include the effects of the magnetic field.

To make the units work out better, the magnetic field $\mathbf{B}(x, y, z) = \Phi_0/\mathcal{V} \mathbf{f}(x, y, z)$, where $\mathcal{V}$ is the area of the graphene unit cell in Å², $\Phi_0 = h / 2e$ is the magnetic flux quantum, and $\mathbf{f}(x, y, z) = \nabla \times \mathbf{g}(x,y,z)$ is a dimensionless vector function. Note that $\mathbf{f} = 1$ produces a field of about 40000 T.

For the vector potential, we have $\mathbf{A}(x, y, z) =\Phi_0 / \mathcal{V} \mathbf{g}(x, y, z)$, where $\mathbf{g}(x, y, z)$ has the units of Å. Using the definition of the Peierls phase, one gets

\[\phi = -\frac{\pi}{ \Phi_0} \int \mathbf{A}\cdot d\mathbf{l} = -\frac{\pi}{ \mathcal{V}} \int \mathbf{g}(x, y, z) \cdot d\mathbf{l} \rightarrow -\frac{\pi}{ \mathcal{V}} \int \mathbf{g}_{xy}(x, y) \cdot d\mathbf{l}\,.\]

The last step follows from the fact that the graphene system resides in the $xy$ plane, so one needs to retain only the $x$ and $y$ componends of $\mathbf{g}$, as denoted by the subscript $xy$.

Arguments

vec_pot(x,y): $\mathbf{g}_{xy}(x, y)$ in Å with x and y in Å.

a1: GrapheneCoord of the "from" atom.

a2: GrapheneCoord of the "to" atom.

vec_pot(x,y) needs to return a tuple corresponding to the vector potential in $x$ and $y$ directions.

Green's Functions

GrapheneQFT.δG_R — Function

δG_R(z::ComplexF64,
 pairs::Vector{Tuple{GrapheneCoord,GrapheneCoord}},
 s::GrapheneSystem)

The correction to the real-space graphene Green's function in the presence of defects as a function of complex energy z.

The function returns a vector of ComplexF64 for each GrapheneCoord in pairs.

Arguments

z: complex energy

pairs: pairs of GrapheneCoord's for which δG_R is calculated

s: GrapheneSystem for which δG_R is calculated

GrapheneQFT.G_R — Function

G_R(z::ComplexF64,
pairs::Vector{Tuple{GrapheneCoord,GrapheneCoord}},
s::GrapheneSystem)

The full real-space graphene Green's function in the presence of defects as a function of complex energy z.

Arguments

z: complex energy

pairs: pairs of GrapheneCoord's for which G_R is calculated

s: GrapheneSystem for which G_R is calculated

GrapheneQFT.δΓ — Function

δΓ(z::ComplexF64, s::GrapheneSystem)

The correction to the impurity Green's function due to the impurities' interaction with graphene.

Arguments

z: complex energy

s: GrapheneSystem for which δΓ is calculated

GrapheneQFT.Γ — Function

Γ(z::ComplexF64, s::GrapheneSystem)

The full impurity Green's function with the correction due to the impurities' interaction with graphene.

Arguments

z: complex energy

s: GrapheneSystem for which Γ is calculated

Carbon Orbitals

GrapheneQFT.Ψ_pz — Function

Ψ_pz(r::Float64, θ::Float64)

2pz orbital in spherical coordinates.

The radial part is extracted from DFT, while the angular portion is the spherical harmonic for 2pz (n=2, l=1, m=0) is cos(θ). The wave function does not depend on ϕ, as expected for a pz orbital.

Arguments

r: radial distance from the center of the orbital in Bohr radii

θ: polar angle

Note that the function returns 0 for r > 40 since the radial component at r = 40 is ≈4.5×10⁻¹¹.

GrapheneQFT.coulomb_potential_pz — Function

coulomb_potential_pz(R::Float64, τ::Float64)

Coulomb potential generated by a 2pz orbital at distance R and polar angle τ from the center of the orbital.

GrapheneQFT.coulomb_potential_pz_interp — Function

coulomb_potential_pz_interp(R::Float64, τ::Float64)

Interpolated version of coulomb_potential_pz.

Since calculating coulomb_potential_pz involves computing a 3D integral, multiple calculations can be costly. Therefore, this function uses a precomputed grid to approximate coulomb_potential_pz for scenarios when the direct calculation becomes not feasible.

GrapheneQFT.coulomb_energy_pz_pz — Function

coulomb_energy_pz_pz(ρ::Tuple{Float64, Float64, Float64})

Coulomb interaction between two graphene orbitals.

Arguments

ρ: a 3-tuple separation vector (x,y,z) between the orbital centers